feat: 支持自定义 geometry label 布局
